Move sync data structures from downloads manager to another module
